feat: 지출 전체에서 N차를 계산할 수 있도록 인덱스 추가
